+++ library/devices/fellow-stagg-ekg/internal/cli/root.go
@@ -12,7 +12,7 @@ import (
 	"github.com/spf13/cobra"
 )
 
-var version = "2026.6.1"
+var version = "2026.6.2"
 
 type config struct {
 	baseURL string